app/cms/scavenger/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyascii==0.3.0
    # via wagtail
asgiref==3.5.0
    # via django
beautifulsoup4==4.9.3
    # via wagtail
boto3==1.21.23
    # via -r requirements.in
botocore==1.24.23
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==37.0.4
    # via django-encrypted-model-fields
django==4.0
    # via
    #   -r requirements.in
    #   django-encrypted-model-fields
    #   django-filter
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   instagram-sync
    #   wagtail
django-encrypted-model-fields==0.6.5
    # via instagram-sync
django-filter==21.1
    # via wagtail
django-modelcluster==5.3
    # via wagtail
django-storages==1.12.3
    # via -r requirements.in
django-taggit==2.1.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
djangorestframework==3.13.1
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
idna==3.3
    # via requests
instagram-sync[django]==0.1.7
    # via -r requirements.in
jinja2==3.0.3
    # via -r requirements.in
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
jsonformatter==0.3.1
    # via -r requirements.in
l18n==2021.3
    # via wagtail
markupsafe==2.1.1
    # via jinja2
openpyxl==3.0.9
    # via tablib
pendulum==2.1.2
    # via -r requirements.in
pillow==9.0.1
    # via wagtail
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via
    #   botocore
    #   pendulum
pytz==2022.1
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pytzdata==2020.1
    # via pendulum
requests==2.28.1
    # via
    #   instagram-sync
    #   wagtail
s3transfer==0.5.2
    # via boto3
sentry-sdk==1.5.8
    # via -r requirements.in
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tablib[xls,xlsx]==3.2.0
    # via wagtail
telepath==0.2
    # via wagtail
urllib3==1.26.9
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wagtail==2.16
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via wagtail
xlwt==1.3.0
    # via tablib
